virtio-serial: boot guest with multiple pcis with multiple ports
