+++ qemu/tests/cfg/virtio_serial_file_transfer.cfg
@@ -30,14 +30,22 @@
                 - from_guest_to_host:
                     file_sender = guest
         - boot_with_multiple_virtserialports:
+            type = boot_N_M_virtserialports
+            only unix_socket
+            filesize = 512
             variants:
                 - 1_N:
-                    only unix_socket
-                    type = boot_1_N_virtserialports
-                    driver_name = vioser
                     serials += " vs2 vs3"
                     serial_type_vs2 = virtserialport
                     serial_type_vs3 = virtserialport
+                - N_M:
+                    serials += " vs2 vs3 vs4 vs5 vs6"
+                    serial_type_vs2 = virtserialport
+                    serial_type_vs3 = virtserialport
+                    serial_type_vs4 = virtserialport
+                    serial_reuse_bus_vs4 = no
+                    serial_type_vs5 = virtserialport
+                    serial_type_vs6 = virtserialport
         - max_ports:
             only unix_socket
             type = virtio_serial_file_transfer_max_ports
